Install @faker-js/faker
This will be used to assign random values to objects created in tests

Also add babel transform work around required to make the latest version of
`@faker/js` work with Jest.

The recent release of Faker v10 introduced a breaking change by removing
CommonJS support entirely, making it ESM-only. This breaks compatibility with
many existing TypeScript/JavaScript projects that use Jest and
other testing frameworks in CommonJS environments:
https://github.com/faker-js/faker/issues/3606#issuecomment-3233612736

In order to use the faker-js faker library with Jest, we will either need to
pin the version to v9.9, or use the following workaround:

- Install `babel-jest` and `@babel/preset-env`
- Use `babel-jest` to transform JavaScript files, with a preset setup (using
  `preset-env`)
- Ensure that the package in question is not transformed by ts-jest by adding a
  `transformIgnorePattern`.