Add Dto Faker factories
These are used to create instances of objects (Dtos) defined by the Community
Payback API for use in tests. It is better to always use one of these factories
to setup test data as it reduces the number of breaking changes when Dto models
change.

These are copied over from Factories defined in the hmpps-community-payback-ui
repository. I have copied them all with the exception of those for Dtos which
are not currently defined on the schema for the Supervisor UI APIs.