Add session details to session page
This will include the key details about a session, including project name, date and location.

I have used html rather than the nunjucks template for the use of the GOV.UK
notification banner, as I feel this has been used for styling only - and we
don't want it to have an alert role or associated JS.

I have left the address out of integration tests as the Faker data could have
null values, which makes the result inconsistent.