Show list of offenders
Note - Using html for the summary list component rather than nunjucks:

The styling uses that of the GOV.UK summary list, but as our cards do not
currently contain description list content it would be incorrect to use the
nunjucks template, as this would create an empty list.

Also we need a h3 heading level, not h2.