Add start time title
As this will vary by form route, moving to the page model.

Add start of integration test. I will expand this test to validate the form,
but for now adding the visit part of the test, which will check the page title
is correct. I am anticipating we will use this page model for both 'Arrival'
and 'Absent' forms, so setting two possible titles based on whether it is an
arrival form. We will need to add a different visit path for an absent form.