Handle absent action in start time view
This will be managed by an `AppointmentArrivedAction` type passed from the
route handler to controller and page model.
If the path is for an 'absent' journey, return different paths and page title.

Also refactor startTime page to take the new type to calculate the expected
title instead of a boolean. An alternative option for this would be to have two
distinct page models for tests (perhaps with a common base).