Introduce ReferenceDataService
We want to display a subset of contact outcomes depending on the context of the
journey and question. Here we create a method for returning only the contact
outcomes related to a person that has arrived but is unable to work.

This logic may change in the future, so we've implemented this logic in the
service layer to allow minimal disruption if we have to refactor.