Add appointment status service
With method to get or create statuses for a session.

We want to calculate and maintain a list of statuses for a session, so creating
a single form object for each session, which is calculated if it has not been
created previously.

A unique identifier for a session is a combination of the project code and session date,
so using this as the form id to ensure one status list is created per session.

Note: it is likely that we will add futher statuses, including different
starting or 'completed' statuses for appointments with certain contact
outcomes.