Add method to get a single appointment status
As we are storing statuses as a list of statuses by appointment id for a
session, this method will look up the full session status list and look up the
right one for the appointment, but the interface will hide this implementation
should we change this to an API request.

This implementation expects that the session statuses will have already been
set up in a previous journey, and will error if not found.