Temporarily remove e2e assertion for status type
As statuses are persisted for a session, and we currently don't have a workflow
which creates fresh data for the test run, removing this test for the
appointment status, as the test will fail as soon as the status is updated once.

Plan to add a temporary workaround to clear status data to enable us to test the status change and
finish journey in e2e tests. Moving the original assertion to a new unused
method, so we can easily add it in to tests later.