Add util to customise footer links
Any controller can pass a `footerLinks` array to the view to add custom links
to the front of the footer links list.

We wouldn't normally customise footer links per page, but this allows us to add
a 'Clear session data' link to only the session page for dev environments, but
keep it separate from the normal UI to prevent confusion.