Add feature flag for clearing session data
We want to enable a link to clear saved session status data in development
environments, as we are currently testing with one session and we want to reset
the appointments to enable repeated testing. We never want to expose this in
production environments so it will be hidden behind a feature flag.