Save contact outcome with compliance data
If the appointment is finished - on time ('completed') action, then we should
save the 'attended - complied' contact outcome.

Storing this code on the reference data service, but not performing a lookup.

Assuming that if not 'completed' journey we can pass the previously selected
contact outcome to the `requestData` method instead.