Introduce LeftEarlyReasonController and route handlers
Note, we're calling the same method on the referenceData service for retrieving
outcomes as at present, these are the same as the unable to work page. These
will likely change in the future.

This method name will be updated in a follow up piece of work to make it
suitable for both arrived - unable to work and arrived - left early journey
flows.