Remove creation of default session statuses on `getStatusesForSession`
Previously, we were creating a status form object for a session at the point of
fetching all statuses (on a session list page). This meant we would have to
handle creating and saving a default status for any new appointments, and
changing the default status if the appointment outcome changed before the
appointment reaches an active state recognised by our interface.

Instead, we will now only persist a status once it reaches an active status in
our UI - and simply display a default status if one doesn't exist. This
approach gives us much more flexibility.