Skip arrived - unable to work e2e
Our implementation of this journey does not currently work as expected because
we are asking the supervisor to select a contact outcome which requires
`attendanceData` in order to save successfully to Delius.

The test is passing because our mock data has this data already populated. It
would be good to clear this and test a more realistic scenario - so skipping
the e2e until the bug is fixed.