Ensure form is populated with entered answers if errors
If a user attempts to submit the form but there are errors, then they should
see the page re-rendered with their submitted answers and any errors.

To reduce inline logic when formatting viewData for inputs, I have created one
form values object with the `ComplianceQuery` interface. As the query value for
yes/no radios is of type `yes` or `no` rather than a boolean, I have changed
the type the `GovUkRadio.yesNoItems` takes, and instead always pass in this
type. We can cast the appointment boolean values outside of the method.

As still populating with appointment values if no query, adding tests for this.