Update next session service method to fetch allocations for supervisor
Previously, we fetched 'next sessions' for a hardcoded provider and team, but
now we are fetching information about a supervisor's unpaid work team we can
now replace this.

As a supervisor can be on multiple unpaid work teams, and the next session
endpoint only takes details for one team, we need to handle the possibility of
multiples. We have opted to make a request for each team and join them into one
list.

I have kept this logic to the service method, particularly as this feels
like a temporary solution.

This update requires quite a few updates to integration tests, as this request
happens on the index page, which is visited every time we call `cy.signIn()`.
It may be worth passing a url to this method to prevent repeated stub mapping
for the homepage when the test is testing visiting a different page.

Also, update team used in e2e tests to the supervisor Test User team, to ensure
appointments are created for the right team for our test user.