Require isUnpaidWorkTeamMember boolean to be true to access the app
Utilising the new SupervisorService/getSupervisor endpoint, we can now
determine that a user is an unpaid work team member, which we check
directly in the authorisation middleware. If they aren't, we show the
usual auth error page.

Because I was injecting dataAccess() into the authorisation middleware,
this caused a slight problem with the way ApplicationInfo was trying to
retrieve package.json - I switched this from a relative path to relying
on the CWD and that fixed it; here's hoping this doesn't cause any
problems elsewhere.