Replace hardcoded index for delius data
We create only the number of appointments required for the number of tests
being run, using the `CountReporter`. However, if a developer wants to run only
one test (which is not the first), this will fail sue to the hard coded index
used in the reading of delius data. Instead, using the `parallelIndex` property
on the test info.